Tried without success to use the AstFullTraverser (aft) class.
The immediate problem is that aft.visit does not accept a parent argument,
and attempting to "simulate" the argument is a big change to program logic.

The only obvious approach would be to cut/paste all aft members, and add parent args to all methods.
This might introduce other problems, all in the attempt to avoid computing fields.
So the whole project seems a bit dubious...